Non-simple multi-components module.
